stream-processor/requirements.txt
nats-py==2.10.0
pyspark==3.5.5
